Handled error and cleaned up code architecture.
